refactored structure and fixed Dates bug
