+++ src/extract.jl
@@ -76,9 +76,9 @@ function load(file::String, vari::String; poly = ([]), start_date::Tuple=(Inf,),
 
   timeattrib = Dict(ds["time"].attrib)
   timeV = ds["time"][:]
-  f = typeof(timeV[1])
+  T = typeof(timeV[1])
 
-  idxtimebeg, idxtimeend = ClimateTools.timeindex(timeV, start_date, end_date, frequency, f)
+  idxtimebeg, idxtimeend = ClimateTools.timeindex(timeV, start_date, end_date, T)
 
   timeV = timeV[idxtimebeg:idxtimeend]
 
@@ -526,301 +526,6 @@ function load2D(file::String, vari::String; poly=[], data_units::String="")
 
 end
 
-"""
-    shapefile_coords(poly::Shapefile.Polygon)
-
-This function return the polygons contained in shp.shapes[i]. It returns the x and y coordinates vectors.
-
-See also [`shapefile_coords_poly`](@ref), which returns a polygon that ca be used for data extraction of the [`load`](@ref).
-
-"""
-function shapefile_coords(poly::Shapefile.Polygon)
-    start_indices = poly.parts .+ 1
-    end_indices = vcat(poly.parts[2:end], length(poly.points))
-    x, y = zeros(0), zeros(0)
-    for (si,ei) in zip(start_indices, end_indices)
-        push!(x, NaN)
-        push!(y, NaN)
-        for pt in poly.points[si:ei]
-            push!(x, pt.x)
-            push!(y, pt.y)
-        end
-    end
-    x, y
-end
-
-
-"""
-    shapefile_coords_poly(poly::Shapefile.Polygon)
-
-Return the polygons contained in shp.shapes[i]. It returns an array containing the polygons.
-
-See also [`shapefile_coords`](@ref), which returns vectors as opposed to array. Returned polygon is consistent with the data extraction of the [`load`](@ref) function.
-
-"""
-function shapefile_coords_poly(poly::Shapefile.Polygon)
-    x, y = shapefile_coords(poly)
-    return [x y]'
-end
-
-"""
-    extractpoly(file::String, n::Int)
-
-Returns the n-th polygon contained in *file*.
-"""
-function extractpoly(file::String, n::Int)
-
-    shp = open(file) do fd
-        read(fd, Shapefile.Handle)
-    end
-
-    poly = shapefile_coords_poly(shp.shapes[n])
-    return poly
-end
-
-"""
-    timeresolution(timevec::Array{N,1} where N)
-
-Return the time resolution of the vector timevec.
-
-"""
-function timeresolution(timevec::Array{N,1} where N)
-
-    # timevec = (NetCDF.ncread(str, "time"))
-    if length(timevec) > 1
-        timediff = diff(timevec)[2]
-        if timediff == 1. || timediff == 1
-            return "24h"
-        elseif round(timediff, digits=5) == round(12/24, digits=5)
-            return "12h"
-        elseif round(timediff, digits=5) == round(6/24, digits=5)
-            return "6h"
-        elseif round(timediff, digits=5) == round(3/24, digits=5)
-            return "3h"
-        elseif round(timediff, digits=5) == round(1/24, digits=5)
-            return "1h"
-        elseif round(timediff, digits=5) == 365.0 || round(timediff, digits=5) == 366.0 || round(timediff, digits=5) == 360.0
-            return "Yearly"
-        end
-    else
-        return "N/A"
-    end
-end
-
-"""
-    function pr_timefactor(rez::String)
-
-Return the time factor that should be applied to precipitation to get accumulation for resolution "rez"
-
-"""
-function pr_timefactor(rez::String)
-
-    if rez == "24h"
-        return 86400.0
-    elseif rez == "12h"
-        return 43200.0
-    elseif rez == "6h"
-        return 21600.0
-    elseif rez == "3h"
-        return 10800.0
-    elseif rez == "1h"
-        return 3600.0
-    elseif rez == "N/A"
-        return 1.0
-    end
-
-end
-
-"""
-    function daymean_factor(rez::String)
-
-Return the time factor that should be applied to precipitation to get accumulation for resolution "rez"
-
-"""
-function daymean_factor(rez::String)
-
-    if rez == "24h" || rez == "day" || rez == "daily"
-        return 1
-    elseif rez == "12h"
-        return 2
-    elseif rez == "6h"
-        return 4
-    elseif rez == "3h"
-        return 8
-    elseif rez == "1h"
-        return 24
-    elseif rez == "N/A"
-        return 1
-    end
-
-end
-
-"""
-    spatialsubset(C::ClimGrid, poly::Array{N, 2})
-
-Returns the spatial subset of ClimGrid C. The spatial subset is defined by the polygon poly, defined on a -180, +180 longitude reference.
-
-"""
-function spatialsubset(C::ClimGrid, poly)
-
-    # Some checks for polygon poly
-
-    if size(poly, 1) != 2 && size(poly, 2) == 2
-        # transpose
-        poly = poly'
-    end
-
-
-    lonsymbol = Symbol(C.dimension_dict["lon"])
-    latsymbol = Symbol(C.dimension_dict["lat"])
-
-    longrid = C.longrid
-    latgrid = C.latgrid
-
-    lon = C[1][Axis{lonsymbol}][:]
-    lat = C[1][Axis{latsymbol}][:]
-
-    msk = inpolygrid(longrid, latgrid, poly)
-    # idlon, idlat = findn(.!isnan.(msk))
-    begin
-        I = findall(!isnan, msk)
-        idlon, idlat = (getindex.(I, 1), getindex.(I, 2))
-    end
-    minXgrid = minimum(idlon)
-    maxXgrid = maximum(idlon)
-    minYgrid = minimum(idlat)
-    maxYgrid = maximum(idlat)
-
-    # Get DATA
-    data = C[1].data
-
-    if ndims(data) == 3
-        data = data[minXgrid:maxXgrid, minYgrid:maxYgrid, :]
-    elseif ndims(data) == 4
-        data = data[minXgrid:maxXgrid, minYgrid:maxYgrid, :, :]
-    end
-
-    #new mask (e.g. representing the region of the polygon)
-    msk = msk[minXgrid:maxXgrid, minYgrid:maxYgrid]
-    data = applymask(data, msk)
-
-    # Get lon-lat for such region
-    longrid = longrid[minXgrid:maxXgrid, minYgrid:maxYgrid]
-    latgrid = latgrid[minXgrid:maxXgrid, minYgrid:maxYgrid]
-    lon = lon[minXgrid:maxXgrid]
-    lat = lat[minYgrid:maxYgrid]
-
-    if ndims(data) == 3
-      dataOut = AxisArray(data, Axis{lonsymbol}(lon), Axis{latsymbol}(lat),  Axis{:time}(C[1][Axis{:time}][:]))
-    elseif ndims(data) == 4
-      dataOut = AxisArray(data, Axis{lonsymbol}(lon), Axis{latsymbol}(lat), Axis{:level}(C[1][Axis{:plev}][:]), Axis{:time}(C[1][Axis{:time}][:]))
-    end
-
-    return ClimGrid(dataOut, longrid=longrid, latgrid=latgrid, msk=msk, grid_mapping=C.grid_mapping, dimension_dict=C.dimension_dict, model=C.model, frequency=C.frequency, experiment=C.experiment, run=C.run, project=C.project, institute=C.institute, filename=C.filename, dataunits=C.dataunits, latunits=C.latunits, lonunits=C.lonunits, variable=C.variable, typeofvar=C.typeofvar, typeofcal=C.typeofcal, varattribs=C.varattribs, globalattribs=C.globalattribs)
-
-end
-
-"""
-    function temporalsubset(C::ClimGrid, startdate::Date, enddate::Date)
-
-Returns the temporal subset of ClimGrid C. The temporal subset is defined by a start and end date.
-
-"""
-function temporalsubset(C::ClimGrid, datebeg::Tuple, dateend::Tuple)
-
-    f = typeof(get_timevec(C)[1])
-
-    startdate = ClimateTools.buildtimetype(datebeg, f)
-    enddate = ClimateTools.buildtimetype(dateend, f)
-
-    # some checkups
-    @argcheck startdate <= enddate
-    @argcheck startdate >= C[1][Axis{:time}][1]
-    @argcheck enddate <= C[1][Axis{:time}][end]
-
-    # Temporal subset
-    dataOut = C[1][Axis{:time}(startdate .. enddate)]
-
-    # The following control ensure that a 1-timestep temporal subset returns a 3D Array with time information on the timestep. i.e. startdate == enddate
-    if ndims(dataOut) == 2
-        timeV = startdate
-        latsymbol = Symbol(C.dimension_dict["lat"])
-        lonsymbol = Symbol(C.dimension_dict["lon"])
-        data2 = fill(NaN, (size(dataOut,1), size(dataOut, 2), 1))
-        data2[:,:,1] = dataOut
-        dataOut = AxisArray(data2, Axis{lonsymbol}(C[1][Axis{lonsymbol}][:]), Axis{latsymbol}(C[1][Axis{latsymbol}][:]), Axis{:time}(C[1][Axis{:time}][:]))
-
-    end
-
-    return ClimGrid(dataOut, longrid=C.longrid, latgrid=C.latgrid, msk=C.msk, grid_mapping=C.grid_mapping, dimension_dict=C.dimension_dict, timeattrib=C.timeattrib, model=C.model, frequency=C.frequency, experiment=C.experiment, run=C.run, project=C.project, institute=C.institute, filename=C.filename, dataunits=C.dataunits, latunits=C.latunits, lonunits=C.lonunits, variable=C.variable, typeofvar=C.typeofvar, typeofcal=C.typeofcal, varattribs=C.varattribs, globalattribs=C.globalattribs)
-
-end
-
-"""
-    periodsubset(C::ClimGrid, startmonth::Int64, endmonth::Int64)
-
-Return the temporal subset of ClimGrid C based on months.
-"""
-function periodsubset(C::ClimGrid, startmonth::Int64, endmonth::Int64)
-
-    @argcheck startmonth >= minimum(Dates.month.(C[1][Axis{:time}][:]))
-    @argcheck startmonth <= maximum(Dates.month.(C[1][Axis{:time}][:]))
-
-    if startmonth <= endmonth
-        # Each matrix [:,:,i] represent data for a day
-        datain = C.data.data
-        # Date vector
-        datevecin = C[1][Axis{:time}][:]
-        # Where are the data between startmonth and endmonth
-        index = (Dates.month.(datevecin) .<= endmonth) .&  (Dates.month.(datevecin) .>= startmonth)
-        # Keep only data between startmonth and endmonth
-        dataout = datain[:,:,index]
-        datevecout = datevecin[index]
-        # Create the ClimGrid output
-        lonsymbol = Symbol(C.dimension_dict["lon"])
-        latsymbol = Symbol(C.dimension_dict["lat"])
-        axisout = AxisArray(dataout, Axis{lonsymbol}(C[1][Axis{lonsymbol}][:]), Axis{latsymbol}(C[1][Axis{latsymbol}][:]),    Axis{:time}(datevecout))
-
-    elseif endmonth < startmonth
-        # Each matrix [:,:,i] represent data for a day
-        datain = C.data.data
-        # Date vector
-        datevecin = C[1][Axis{:time}][:]
-        # Where are the data between startmonth and endmonth
-        index = (Dates.month.(datevecin) .<= endmonth) .|  (Dates.month.(datevecin) .>= startmonth)
-        # Keep only data between startmonth and endmonth
-        dataout = datain[:,:,index]
-        datevecout = datevecin[index]
-        # Create the ClimGrid output
-        lonsymbol = Symbol(C.dimension_dict["lon"])
-        latsymbol = Symbol(C.dimension_dict["lat"])
-        axisout = AxisArray(dataout, Axis{lonsymbol}(C[1][Axis{lonsymbol}][:]), Axis{latsymbol}(C[1][Axis{latsymbol}][:]), Axis{:time}(datevecout))
-    end
-
-    return ClimGrid(axisout, longrid=C.longrid, latgrid=C.latgrid, msk=C.msk, grid_mapping=C.grid_mapping,dimension_dict=C.dimension_dict, timeattrib=C.timeattrib, model=C.model, frequency=C.frequency, experiment=C.experiment, run=C.run, project=C.project,institute=C.institute, filename=C.filename, dataunits=C.dataunits, latunits=C.latunits, lonunits=C.lonunits, variable=C.variable,typeofvar=C.typeofvar, typeofcal=C.typeofcal, varattribs=C.varattribs, globalattribs=C.globalattribs)
-end
-
-"""
-    periodsubset(C::ClimGrid, season::String)
-
-Return the temporal subset of ClimGrid C for a given season. Season options are: "DJF" (December-February), "MAM" (March-May), "JJA" (June-August), "SON" (September-November)
-"""
-function periodsubset(C::ClimGrid, season::String)
-    if lowercase(season) == "djf"
-      D = periodsubset(C, 12, 2)
-    elseif lowercase(season) == "mam"
-      D = periodsubset(C, 3, 5)
-    elseif lowercase(season) == "jja"
-      D = periodsubset(C, 6, 8)
-    elseif lowercase(season) == "son"
-      D = periodsubset(C, 9, 11)
-    else
-      error("Wrong season name. Options are DJF (December-February; Winter), MAM (March-May; Spring), JJA (June-August; Summer) and SON (September-November; Fall)")
-    end
-
-    return D
-end
-
 model_id(attrib::NCDatasets.Attributes) = get(attrib,"model_id",get(attrib,"model","N/A"))
 
 experiment_id(attrib::NCDatasets.Attributes) = get(attrib,"experiment_id",get(attrib,"experiment","N/A"))
@@ -866,98 +571,6 @@ function getdim_lon(ds::NCDatasets.Dataset)
 
 end
 
-function shiftgrid_180_west_east(longrid)
-
-    for ilon in eachindex(longrid)
-        if longrid[ilon] >= 180
-            longrid[ilon] = longrid[ilon] - 360
-        end
-    end
-
-    ieast = longrid .>= 0
-    iwest = longrid .< 0
-
-    grideast = reshape(longrid[ieast], :, size(longrid, 2))
-    gridwest = reshape(longrid[iwest], :, size(longrid, 2))
-
-    longrid_flip = vcat(gridwest, grideast)
-
-    return longrid_flip
-
-end
-
-
-function shiftgrid_180_east_west(longrid)
-
-    for ilon in eachindex(longrid)
-        if longrid[ilon] >= 180
-            longrid[ilon] = longrid[ilon] - 360
-        end
-    end
-
-    ieast = longrid .>= 0
-    iwest = longrid .< 0
-
-    grideast = reshape(longrid[ieast], :, size(longrid, 2))
-    gridwest = reshape(longrid[iwest], :, size(longrid, 2))
-
-    longrid_flip = vcat(grideast, gridwest)
-
-    return longrid_flip
-
-end
-
-function shiftarray_west_east(data, longrid_flip)
-
-    ieast = longrid_flip .>= 0
-    iwest = longrid_flip .< 0
-
-    msk = permute_west_east2D(data, iwest, ieast)
-    return msk
-
-end
-
-function shiftarray_east_west(data, longrid_flip)
-
-    ieast = longrid_flip .>= 0
-    iwest = longrid_flip .< 0
-
-    msk = permute_east_west2D(data, iwest, ieast)
-    return msk
-
-end
-
-
-function shiftvector_180_east_west(lon_raw)
-
-    for ilon in eachindex(lon_raw)
-        if lon_raw[ilon] >= 180
-            lon_raw[ilon] = lon_raw[ilon] - 360
-        end
-    end
-    ieast_vec = lon_raw .>= 0
-    iwest_vec = lon_raw .< 0
-    lon_raw_flip = [lon_raw[ieast_vec];lon_raw[iwest_vec]]
-    return lon_raw_flip
-
-end
-
-function shiftvector_180_west_east(lon_raw)
-
-    for ilon in eachindex(lon_raw)
-        if lon_raw[ilon] >= 180
-            lon_raw[ilon] = lon_raw[ilon] - 360
-        end
-    end
-
-    ieast_vec = lon_raw .>= 0
-    iwest_vec = lon_raw .< 0
-
-    lon_raw_flip = [lon_raw[iwest_vec];lon_raw[ieast_vec]]
-
-    return lon_raw_flip
-
-end
 
 """
     extractdata(data, msk, idxtimebeg, idxtimeend)
@@ -1005,19 +618,5 @@ function extractdata2D(data, msk)
     data = data[minXgrid:maxXgrid, minYgrid:maxYgrid]
 
     return data
-
-end
-
-function meridian_check(poly)
-
-    minpoly = minimum(poly[1, .!isnan.(poly[1, :])])
-    maxpoly = maximum(poly[1, .!isnan.(poly[1, :])])
-
-    meridian = false
-    if sign(minpoly)*sign(maxpoly) == sign(-1.0) # polygon crosses the meridian
-       meridian = true
-    end
-
-    return meridian
 
 end
