+++ src/interface.jl
@@ -183,79 +183,3 @@ Base.show(io::IO, ::MIME"text/plain", ITP::TransferFunction) = print(io, "Transf
     "Interpolation array: ", size(ITP.itp), " transfer functions", "\n",
     "Method: ", ITP.method, "\n",
     "Detrended: ", ITP.detrend)
-
-
-"""
-    timeindex(timeVec, start_date, end_date, freq, f)
-
-Return the index of time vector specified by start_date and end_date. Provide timestep "freq" to account for monthly timestep. f is the DateTime type (see NCDatasets.jl documentation).
-"""
-function timeindex(timeV, datebeg::Tuple, dateend::Tuple, frequency, f)
-
-    # Start Date
-    if !isinf(datebeg[1])
-
-        # Build DateTime type
-        start_date = ClimateTools.buildtimetype(datebeg, f)
-
-        # Check
-        @argcheck start_date >= timeV[1]
-
-        if frequency != "mon"
-            idxtimebeg = findfirst(timeV .== start_date)[1]
-        elseif frequency == "mon"
-            idxtimebeg = findfirst(timeV .== start_date)[1]
-        end
-    else
-        idxtimebeg = 1
-
-    end
-
-    # End date
-    if !isinf(dateend[1])
-
-        # Build DateTime type
-        end_date = ClimateTools.buildtimetype(dateend, f)
-
-        @argcheck end_date <= timeV[end]
-        if frequency != "mon"
-            idxtimeend = findlast(timeV .== end_date)[1]
-        elseif frequency == "mon"
-            idxtimeend = findlast(timeV .== Date(Dates.year(end_date), Dates.month(end_date), Dates.day(1)))[1]
-        end
-    else
-        idxtimeend = length(timeV)
-    end
-
-    if !isinf(datebeg[1]) && !isinf(dateend[1])
-        @argcheck start_date <= end_date
-    end
-
-    return idxtimebeg, idxtimeend
-end
-
-"""
-    buildtimetype(datetuple, f)
-
-Returns the adequate DateTime for temporal subsetting using DateType *f*
-"""
-function buildtimetype(date_tuple, f)
-
-    if length(date_tuple) == 1
-        dateout = f(date_tuple[1], 01, 01)
-    elseif length(date_tuple) == 2
-        dateout = f(date_tuple[1], date_tuple[2], 01)
-    elseif length(date_tuple) == 3
-        dateout = f(date_tuple[1], date_tuple[2], date_tuple[3])
-    elseif length(date_tuple) == 4
-        dateout = f(date_tuple[1], date_tuple[2], date_tuple[3], date_tuple[4], 00, 00)
-    elseif length(date_tuple) == 5
-        dateout = f(date_tuple[1], date_tuple[2], date_tuple[3], date_tuple[4], date_tuple[5], 00)
-    elseif length(date_tuple) == 6
-        dateout = f(date_tuple[1], date_tuple[2], date_tuple[3], date_tuple[4], date_tuple[5], date_tuple[6])
-    end
-
-    return dateout
-
-
-end
